Replace connector status classes with generic statuses
