+++ application/test/connectors/connector_class_dispatcher_test.rb
@@ -3,18 +3,6 @@ require 'test_helper'
 
 class ConnectorClassDispatcherTest < ActiveSupport::TestCase
 
-  test 'download_connector_status should return Dataverse::DownloadConnectorStatus class for dataverse files' do
-    project = download_project(type: ConnectorType::DATAVERSE, files: 1)
-    result = ConnectorClassDispatcher.download_connector_status(project.download_files.first)
-    assert_instance_of Dataverse::DownloadConnectorStatus, result
-  end
-
-  test 'upload_file_connector_status should return Dataverse::UploadConnectorStatus class for dataverse files' do
-    project = upload_project(type: ConnectorType::DATAVERSE, files: 1)
-    result = ConnectorClassDispatcher.upload_file_connector_status(project.upload_bundles.first, project.upload_bundles.first.files.first)
-    assert_instance_of Dataverse::UploadConnectorStatus, result
-  end
-
   test 'download_connector_metadata should return Dataverse::DownloadConnectorMetadata class for dataverse files' do
     project = download_project(type: ConnectorType::DATAVERSE, files: 1)
     result = ConnectorClassDispatcher.download_connector_metadata(project.download_files.first)
@@ -59,7 +47,7 @@ class ConnectorClassDispatcherTest < ActiveSupport::TestCase
   test 'raises ConnectorNotSupported for unknown connector type' do
     file = OpenStruct.new(type: :unknown)
     error = assert_raises(ConnectorClassDispatcher::ConnectorNotSupported) do
-      ConnectorClassDispatcher.download_connector_status(file)
+      ConnectorClassDispatcher.download_connector_metadata(file)
     end
     assert_match /Invalid connector type/, error.message
   end
