--- application/test/fixtures/zenodo/download_connector_status/100bytes_file.txt.part
@@ -1,1 +1,0 @@
-ZQXLRKPBJAGVDMHTNCWYFUOEISZQXLRKPBJAGVDMHTNCWYFUOEISZQXLRKPBJAGVDMHTNCWYFUOEISZQXLRKPBJAGVDMHTNCWYFU
