--- application/test/fixtures/zenodo/download_connector_status/completed_file.txt
@@ -1,1 +1,0 @@
-ZQXLRKPBJAGVDMHTNCWYFUOEISZQXLRKPBJAGVDMHTNCWYFUOEISZQXLRKPBJAGVDMHTNCWYFUOEISZQXLRKPBJAGVDMHTNCWYFU
